move TokenStore into its own file
and move its specs into their own file
